pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-cxo"
version = "0.1.0"
description = "Agentic CXO Context Pipeline — AI-driven C-suite agents with modular context management"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Agentic CXO Team"}]
keywords = ["ai", "agents", "cxo", "rag", "context-pipeline", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai>=1.14",
    "tiktoken>=0.7",
    "numpy>=1.26",
    "chromadb>=0.5",
    "pydantic>=2.6",
    "fastapi>=0.111",
    "uvicorn>=0.29",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "rich>=13.7",
    "typer>=0.12",
    "PyPDF2>=3.0",
    "python-docx>=1.1",
    "python-dotenv>=1.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "cryptography>=42.0",
    "apscheduler>=3.10",
    "sse-starlette>=2.0",
    "sqlalchemy>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[project.scripts]
cxo = "agentic_cxo.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"src/agentic_cxo/scenarios/analyst.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
